Updates for NCI

+++ lxs599/umplot/conv_um2nc.sh
@@ -2,38 +2,52 @@
 
 ######################################################################
 
-# Cherax - ste69f, NCI - lxs599 ----------------------
-# Cherax - dix043, NCI - mrd599 ----------------------
-if ( $HOSTNAME == cherax ) then
+# Ruby - ste69f, NCI - lxs599 ----------------------
+# Ruby - dix043, NCI - mrd599 ----------------------
+if ( $HOSTNAME == ruby ) then
  set CONV2NC   = ~ste69f/umutils/conv2nc.tcl
+ #set CONV2NC4  = 
  set CONV2NCTS = ~dix043/src/python/um/um_timeseries.py
  set FLDSUBSET = ~dix043/src/python/um/um_fields_subset.py
 else
  set CONV2NC   = ~lxs599/umutils/conv2nc.tcl
+ set CONV2NC4  = ~access/bin/um2netcdf4.py
  set CONV2NCTS = ~mrd599/src/python/um/um_timeseries.py
  set FLDSUBSET = ~mrd599/src/python/um/um_fields_subset.py
 endif
 set a = a
+@ nom = ($FullYrs * 12) # number of files for no. of years and months
+@ nod = ($FullYrs * 12 * 31) # number of files for no. of years, months and days
 
 ######################################################################
 
 #    cd $DIR
+    #set extlist = 'pm pa pe pb pc pd pf pg ph ps py px'
     set extlist = 'pm pa pe pb pc pi pj pd pf pg ph ps py px'
     foreach ext ( $extlist )
          
       if ( $CPL == y ) then
-       set fillist = `ls $RUNID.$ext-??????????`
+       set fillist = `ls $RUNID.$ext-?????????? | head -$nom`
        foreach file ( $fillist )
         if (! -e $file.nc) then
          echo "Converting " $file " to " $file".nc"
          $CONV2NC -i $file -o $file.nc
+         #python $CONV2NC4 -i $file -o $file.nc
         else
          echo "File" $file "Already Exists!"
         endif
        end # fe file
       else # CPL
        set i = 1
-       set fillist = `ls $RUNID$a.$ext?????`
+       if ($ext == $Ptimes && $RES == 320) then
+       set fillist = `ls $RUNID$a.$ext????? | head -$nod`
+       else
+       if ($VN == 85) then
+       set fillist = `ls $RUNID$a.$ext??????? | head -$nom`
+       else
+       set fillist = `ls $RUNID$a.$ext????? | head -$nom`
+       endif
+       endif
        #set fillist = `ls $DIR/$RUNID$a.$ext?????`
        set cfiles = ${#fillist}
 
@@ -56,32 +70,57 @@ set a = a
          if (! -e $newlist[$i].nc) then
           echo "Converting " $file " to " $newlist[$i]".nc"
           if ($ext == $Ptimes) then
-           python $FLDSUBSET -i $file -o $file.sub2 -x 1201 -x 1235 -x 2201 -x 2207
+           python $FLDSUBSET -i $file -o $file.sub2 -x 1201 -x 1235 -x 2201 -x 2207 -x 0238 -x 2204 -x 2407
            $CONV2NC -i $file.sub2 -o $newlist[$i]_noswlw.nc
-           python $FLDSUBSET -i $file -o $file.sub3 -v 1201 -v 1235 -v 2201 -v 2207
+           python $FLDSUBSET -i $file -o $file.sub3 -v 1201 -v 1235 -v 2201 -v 2207 -v 0238 -v 2204 -v 2407
            $CONV2NC -i $file.sub3 -o $newlist[$i]_swlw.nc
            #set cdate=`cdo showdate $newlist[$i]_swlw.nc`
            #set ctime=`cdo showtime $newlist[$i]_swlw.nc`
            #cdo inttime,$cdate[1],$ctime[1],30minutes $newlist[$i]_swlw.nc $newlist[$i]_swlw_intp30min.nc
            python $FLDSUBSET -i $file -o $file.sub -v 3234 -v 3236 -v 3217 -v 5226 -v 3333
-           python $CONV2NCTS -i $file.sub -o $newlist[$i].nc
-           #$CONV2NC -i $file.sub -o $newlist[$i].nc
+           #python $CONV2NCTS -i $file.sub -o $newlist[$i].nc
+           $CONV2NC -i $file.sub -o $newlist[$i].nc
            rm $file.sub $file.sub2 $file.sub3
-          else
-           $CONV2NC -i $file -o $newlist[$i].nc
-          endif
-         else
+          else # Ptimes
+           if ($ext == $Pdaily && $Ptemp1 != $Ptemps ) then
+            python $FLDSUBSET -i $file -o $file.sub -v 3236
+            $CONV2NC -i $file.sub -o $newlist[$i].nc
+            #python $CONV2NC4 -i $file.sub -o $newlist[$i].nc
+            rm $file.sub
+           else # Pdaily
+           if ( $TSTEP == 288 && $ext == pa ) then
+            python $CONV2NC4 -i $file -o $newlist[$i].nc
+           else
+            $CONV2NC -i $file -o $newlist[$i].nc
+            #python $CONV2NC4 -i $file -o $newlist[$i].nc
+           endif
+           endif # Pdaily
+          endif # Ptimes
+         else   # newlist
           echo "File" $newlist[$i] "Already Exists!"
-         endif
+         endif  # newlist
          set ntime=`cdo ntime $newlist[$i].nc`
+         if ($VN == 73) then
          if (($ext == pc && $ntime >= 224)) then # 224 = 8ts * 1month
-          python $FLDSUBSET -i $file -o $file.sub2 -x 1201 -x 1235 -x 2201 -x 2207
-          $CONV2NC -i $file.sub2 -o $newlist[$i]_noswlw.nc
-          python $FLDSUBSET -i $file -o $file.sub3 -v 1201 -v 1235 -v 2201 -v 2207
-          $CONV2NC -i $file.sub3 -o $newlist[$i]_swlw.nc
-          #python $FLDSUBSET -i $file -o $file.sub -v 3234 -v 3236 -v 3217 -v 5226 -v 3333
-          #$CONV2NC -i $file.sub -o $newlist[$i].nc
-          rm $file.sub2 $file.sub3 #$file.sub
+          if (! -e $newlist[$i]_swlw.nc) then
+           python $FLDSUBSET -i $file -o $file.sub2 -x 1201 -x 1235 -x 2201 -x 2207 -x 0238 -x 2204 -x 2407
+           $CONV2NC -i $file.sub2 -o $newlist[$i]_noswlw.nc
+           python $FLDSUBSET -i $file -o $file.sub3 -v 1201 -v 1235 -v 2201 -v 2207 -v 0238 -v 2204 -v 2407
+           $CONV2NC -i $file.sub3 -o $newlist[$i]_swlw.nc
+           #python $FLDSUBSET -i $file -o $file.sub -v 3234 -v 3236 -v 3217 -v 5226 -v 3333
+           #$CONV2NC -i $file.sub -o $newlist[$i].nc
+           rm $file.sub2 $file.sub3 #$file.sub
+          endif
+         endif
+         if (($ext == pg || $ext == ph)) then
+          if (! -e $newlist[$i]_swlw.nc) then
+           python $FLDSUBSET -i $file -o $file.sub2 -x 1201 -x 1208 -x 1209 -x 1210 -x 1211 -x 2204 -x 2205 -x 2206 -x 2207 -x 2208
+           $CONV2NC -i $file.sub2 -o $newlist[$i]_noswlw.nc
+           python $FLDSUBSET -i $file -o $file.sub3 -v 1201 -v 1208 -v 1209 -v 1210 -v 1211 -v 2204 -v 2205 -v 2206 -v 2207 -v 2208
+           $CONV2NC -i $file.sub3 -o $newlist[$i]_swlw.nc
+           rm $file.sub2 $file.sub3
+          endif
+         endif
          endif
          @ i++
         end # fe file
